moved each section into components
