+++ src/elements/input.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 let props = props => props;
+
 export const TextField = styled.input.attrs({ type: props.type })`
   border: 4px solid #d1d1d1;
   border-radius: 10px;
